cifs: Improve detect_directory_symlink_target() function
Function detect_directory_symlink_target() is not currently able to detect
if the target path is directory in case the path is in the DELETE_PENDING
state or the user has not granted FILE_READ_ATTRIBUTES permission on the
path. This limitation is written in TODO comment.

Resolve this problem by replacing code which determinate path type by the
query_path_info() callback, which now is able to handle all these cases.

Depends on "cifs: Improve SMB2+ stat() to work also for paths in
DELETE_PENDING state".

Signed-off-by: Pali Rohár <pali@kernel.org>