cifs: Use CREATE_OPTION_EXCLUSIVE when opening file/dir for SMB2+ non-POSIX unlink/rmdir
Using of CREATE_OPTION_EXCLUSIVE against non-POSIX SMB2+ server ensures
that the smb2_remove() function either success and removes the directory
entry or it returns an error that file or directory is in use by other SMB
client and silly rename is required to use.

POSIX-based SMB2+ servers do not have this problems as they should unlink
the directory entry immediately and not transition them into delete pending
state.

This allows the cifs_unlink() and cifs_rmdir() functions against non-POSIX
servers to detect these failures via -EBUSY error from smb2_unlink() and
smb2_rmdir() calls and fallbacks to cifs_rename_pending_delete() which
implements silly rename.

This is the final change which enables the silly rename functionality for
the unlink and rmdir calls in SMB2+ dialects on mounted exports from
Windows servers.

With this change Linux unlink() and rmdir() syscalls called on SMB2+ mounts
from Windows servers cause that on success the path would not exist anymore
and new file or directory with that path can be created.

Signed-off-by: Pali Rohár <pali@kernel.org>